@@ -20,7 +20,7 @@ package com.arp.solitaire.dialogs;
 import android.app.Dialog;
 
 import android.os.Bundle;
-import android.support.v4.app.DialogFragment;
+import androidx.fragment.app.DialogFragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.View.OnClickListener;
